+++ _examples/websocket/chat/main.go
@@ -26,8 +26,13 @@ func main() {
 func setupWebsocket(app *iris.Application) {
 	// create our echo websocket server
 	ws := websocket.New(websocket.Config{
+		// These are low-level optionally fields,
+		// user/client can't see those values.
 		ReadBufferSize:  1024,
 		WriteBufferSize: 1024,
+		// only javascript client-side code has the same rule,
+		// which you serve using the ws.ClientSource (see below).
+		EvtMessagePrefix: []byte("my-custom-prefix:"),
 	})
 	ws.OnConnection(handleConnection)
 
@@ -38,7 +43,7 @@ func setupWebsocket(app *iris.Application) {
 	// serve the javascript built'n client-side library,
 	// see websockets.html script tags, this path is used.
 	app.Any("/iris-ws.js", func(ctx iris.Context) {
-		ctx.Write(websocket.ClientSource)
+		ctx.Write(ws.ClientSource)
 	})
 }